clippy: fix lints surfaced on the bare-metal unikernel build path
The clippy aspect runs on every `bazel build`, but the bare-metal
crates are only clippy-checked under the `x86_64_unikernel` /
`aarch64_unikernel` platforms — which the default `aarch64_macos` dev
path skips as incompatible. A rustc/clippy toolchain bump surfaced
~100 lints across those crates that `-Dwarnings` turns into hard build
errors.

Mechanical, behavior-preserving fixes:
  - missing_safety_doc: real `# Safety` sections on the unsafe fns in
    boot, kernel, and the bus / virtio drivers
  - collapsible_if -> edition-2024 let-chains (gve, kernel eventloop,
    aarch64 fdt / exceptions)
  - needless_range_loop -> iterator forms (gve, virtio-net, idt, ghash)
  - manual_div_ceil -> `u*::div_ceil` (gve, virtio bus)
  - manual_range_contains -> `Range::contains` (boot, idt, acpi, time)
  - unnecessary_cast, single_match, needless_return, redundant_closure,
    unnecessary_map_or, fn_to_numeric_cast, no_effect, type_complexity
  - dhcp's 16-arg `set_fallback_config` -> a `FallbackConfig` struct
  - virtio-net `ZEROED` consts: scoped `#[allow(declare_interior_
    mutable_const)]` (one-shot `ptr::write` initializers)
  - mm: `kfree` marked `unsafe fn`; `init` keeps a scoped `#[allow]`
    (marking it unsafe would ripple into crates/boot)

Verified clippy-clean via `bazel build //...` on both unikernel
platforms; the 28 crate unit tests still pass.

+++ crates/drivers/virtio-net/src/init.rs
@@ -229,7 +229,7 @@ pub(crate) fn init_pci_modern() -> bool {
     // dropped. Leaving ctrl uninitialised in the single-queue case
     // avoids the wedge.
     if has_mq && num_pairs > 1 {
-        let ctrl_qi = (2 * max_pairs) as u16;
+        let ctrl_qi = 2 * max_pairs;
         vpci_select_queue(dev, ctrl_qi);
         let ctrl_qsize = vpci_get_queue_size(dev);
         if ctrl_qsize > 0 {
@@ -612,7 +612,7 @@ pub(crate) fn init_mmio() -> bool {
 
     // Init ctrl VQ if MQ negotiated and num_pairs > 1
     if has_mq && num_pairs > 1 {
-        let ctrl_qi = (2 * max_pairs) as u16;
+        let ctrl_qi = 2 * max_pairs;
         unsafe {
             if !(*ndev())
                 .ctrl_queue
